Move click behavior to metabase-lib (#25879)
