+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/ClickBehaviorSidebarHeader/ClickBehaviorSidebarHeader.tsx
@@ -3,7 +3,6 @@ import { t, jt } from "ttag";
 
 import Icon from "metabase/components/Icon";
 
-import { isTableDisplay } from "metabase/lib/click-behavior";
 import {
   isActionDashCard,
   getActionButtonLabel,
@@ -12,6 +11,7 @@ import {
 import type { DashboardOrderedCard } from "metabase-types/api";
 import type { Column } from "metabase-types/types/Dataset";
 
+import { isTableDisplay } from "metabase/lib/click-behavior";
 import { Heading, SidebarHeader } from "../ClickBehaviorSidebar.styled";
 import {
   ColumnClickBehaviorHeader,
