+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/utils.ts
@@ -1,13 +1,12 @@
 import { t } from "ttag";
 import { getIn } from "icepick";
 
-import { hasActionsMenu } from "metabase/lib/click-behavior";
-
 import type {
   ClickBehaviorType,
   DashboardOrderedCard,
 } from "metabase-types/api";
 import type { Column } from "metabase-types/types/Dataset";
+import { hasActionsMenu } from "metabase/lib/click-behavior";
 import { keyForColumn } from "metabase-lib/lib/queries/utils/dataset";
 
 type ClickBehaviorOption = {
