+++ frontend/src/metabase/dashboard/components/ClickMappings.jsx
@@ -10,12 +10,12 @@ import Select from "metabase/core/components/Select";
 
 import MetabaseSettings from "metabase/lib/settings";
 import { isPivotGroupColumn } from "metabase/lib/data_grid";
-import { getTargetsWithSourceFilters } from "metabase/lib/click-behavior";
 import { GTAPApi } from "metabase/services";
 
 import { loadMetadataForQuery } from "metabase/redux/metadata";
 import { getMetadata } from "metabase/selectors/metadata";
 import { getParameters } from "metabase/dashboard/selectors";
+import { getTargetsWithSourceFilters } from "metabase-lib/lib/parameters/utils/click-behavior";
 import Question from "metabase-lib/lib/Question";
 
 class ClickMappingsInner extends React.Component {
