+++ frontend/src/metabase/lib/formatting/url.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import ExternalLink from "metabase/core/components/ExternalLink";
-import { getDataFromClicked } from "metabase/lib/click-behavior";
+import { getDataFromClicked } from "metabase-lib/lib/parameters/utils/click-behavior";
 import { isURL } from "metabase-lib/lib/types/utils/isa";
 import { renderLinkTextForClick, renderLinkURLForClick } from "./link";
 import { formatValue, getRemappedValue } from "./value";
